requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
Pillow
